requirements.txt
python-magic==0.4.27
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
python-whois==0.8.0
geoip2==4.7.0
jinja2==3.1.2
pdfkit==1.0.0
pyyaml==6.0.1
scikit-learn==1.3.2
numpy
pandas==2.0.3
extract-msg==0.41.1
oletools
tldextract==3.4.4
click==8.1.7
rich==13.7.0
olefile
whois==0.9.3
ipwhois==1.2.0
urllib3==1.26.18
python-dateutil==2.8.2
pypdf2==3.0.1
reportlab==4.0.4
# Additional dependencies for enhanced detection
idna==3.4
langdetect==1.0.9
language-tool-python==2.7.1